+++ Platformio/src/device_appleTV/gui_appleTV.cpp
@@ -1,8 +1,8 @@
 #include <lvgl.h>
 // #include "assets.c"
-#include "tft.h"
-#include "gui.h"
-#include "infrared.h"
+#include "gui_general/gui.h"
+#include "hardware/tft.h"
+#include "hardware/infrared_sender.h"
 
 // LVGL declarations
 LV_IMG_DECLARE(appleTvIcon);
