+++ Platformio/src/device_smarthome/gui_smarthome.h
