+++ Platformio/src/hardware/battery.h
